fix: fixed project card image linking
The project card images would not load on the netlify production preview. Likely because SvelteKit
couldn't resolve the variable image paths passed in as a prop.

To remedy this, I moved the project images from the source directory to the static directory
and referenced them there. This seems to have worked.