fix: fixed 500 error
The page was returning a 500 error because I used the onMount() hook without importing it first. In
addition, sometimes the event listeners would try to load before the document was ready.

I imported onMount() and moved the button event listeners to the +page.svelte file and added them on
mount, when the document should be ready.