refactor: changed section structure
The way in which the sections of the site were structured before allowed the home section to overlap
with the project section while landscape on mobile devices.

I've restructured the way the sections on the page work by abstracting away a Section component.
This allowed me to make the margin, padding, and borders of all sections of the site uniform.
In turn, this simplified the code for the sections significantly and allowed me to focus on
what specifically was causing the sections to overlap.

While landscape on mobile devices no longer causes the sections to overlap, they're still ugly
in this view. Maybe come back to it later.